tests: Merge fetch and ajax adapters tests
From now on, dedicated CI jobs will run with and without jQuery.

This lets us simplify the adapters tests a bit. Indeed, the same tests will run
for fetch (without jQuery) and ajax (with jQuery).
